+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### 2.0.1
 
+- 2024 March (Bastien DOIGNIES):
+    - Added semidiscrete optimal transport 2D 
+
 - 2024 Feb (David Coeurjolly)
     - cmake CPM instead of `FETCH_CONTENT` for dependencies 
 - 2024 Fev (Bastien DOIGNIES):